requirements.txt
 pandas
 plotly_express
 streamlit
 vehicles_us.csv